WIP: blow up ejected c.r.a. code...
