+++ package.json
@@ -43,9 +43,7 @@
     "prop-types": "^15.7.2",
     "pubsub-js": "^1.7.0",
     "rc-tree": "2.1.0",
-    "react": "^16.8.6",
     "react-color": "^2.18.0",
-    "react-dom": "^16.8.6",
     "short-number": "^1.0.6",
     "store": "^2.0.12",
     "tinycolor2": "^1.4.1",
@@ -53,6 +51,10 @@
     "uuid": "^3.3.2",
     "vitessce-grid": "0.0.6",
     "whatwg-fetch": "^3.0.0"
+  },
+  "peerDependencies": {
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6"
   },
   "devDependencies": {
     "@babel/core": "7.8.4",
@@ -95,6 +97,7 @@
     "karma-webpack": "4.0.2",
     "mini-css-extract-plugin": "0.9.0",
     "mocha": "7.1.0",
+    "object.fromentries": "^2.0.2",
     "optimize-css-assets-webpack-plugin": "5.0.3",
     "pnp-webpack-plugin": "1.6.0",
     "postcss-flexbugs-fixes": "4.1.0",
@@ -114,7 +117,6 @@
     "wait-on": "^3.2.0",
     "webpack": "4.41.5",
     "webpack-dev-server": "3.10.2",
-    "webpack-manifest-plugin": "2.2.0",
     "workbox-webpack-plugin": "4.3.1"
   },
   "scripts": {
@@ -124,8 +126,8 @@
     "prepublishOnly": "npm run build-lib",
     "clean": "rm -r build-demo build-lib",
     "start": "node scripts/start-demo.js",
-    "test": "karma start config/karma.config.js --single-run",
-    "test:watch": "karma start config/karma.config.js --auto-watch",
+    "test": "karma start scripts/karma.config.js --single-run",
+    "test:watch": "karma start scripts/karma.config.js --auto-watch",
     "cypress:open": "cypress open",
     "cypress:run": "cypress run"
   },
