+++ scripts/getHttpsConfig.js
