+++ scripts/start-demo.js
@@ -10,9 +10,6 @@
 process.env.BABEL_ENV = 'development';
 process.env.NODE_ENV = 'development';
 
-// Ensure environment variables are read.
-require('../config/env');
-
 const fs = require('fs');
 const chalk = require('react-dev-utils/chalk');
 const webpack = require('webpack');
@@ -25,13 +22,18 @@ const {
   prepareUrls,
 } = require('react-dev-utils/WebpackDevServerUtils');
 const { checkBrowsers } = require('react-dev-utils/browsersHelper');
-const paths = require('../config/paths');
-const configFactory = require('../config/webpack.config');
-const createDevServerConfig = require('../config/webpackDevServer.config');
+const getPublicUrlOrPath = require('react-dev-utils/getPublicUrlOrPath');
+
+const paths = require('./paths');
 const utils = require('./utils');
+const configFactory = require('./webpack.config-demo');
+const createDevServerConfig = require('./webpackDevServer.config');
 
 utils.scriptInit();
 utils.checkRequiredFilesForTarget(paths, true);
+
+const isEnvDevelopment = true;
+const publicUrlOrPath = getPublicUrlOrPath(isEnvDevelopment, ".", process.env.PUBLIC_URL);
 
 const useYarn = false;
 const isInteractive = process.stdout.isTTY;
@@ -56,16 +58,16 @@ checkBrowsers(paths.appPath, isInteractive)
       return;
     }
     
-    const config = configFactory(process.env.NODE_ENV, 'demo');
+    const config = configFactory(paths, process.env.NODE_ENV);
     const protocol = process.env.HTTPS === 'true' ? 'https' : 'http';
     const appName = require(paths.appPackageJson).name;
-    const useTypeScript = fs.existsSync(paths.appTsConfig);
-    const tscCompileOnError = process.env.TSC_COMPILE_ON_ERROR === 'true';
+    const useTypeScript = false;
+    const tscCompileOnError = false;
     const urls = prepareUrls(
       protocol,
       HOST,
       port,
-      paths.publicUrlOrPath.slice(0, -1)
+      publicUrlOrPath.slice(0, -1)
     );
     const devSocket = {
       warnings: warnings =>
@@ -89,7 +91,7 @@ checkBrowsers(paths.appPath, isInteractive)
     const proxyConfig = prepareProxy(
       proxySetting,
       paths.appPublic,
-      paths.publicUrlOrPath
+      publicUrlOrPath
     );
     // Serve webpack assets generated by the compiler over a web server.
     const serverConfig = createDevServerConfig(
