+++ src/components/sets/index.js
@@ -1,0 +1,2 @@
+export { default as CellSetsManagerSubscriber } from './CellSetsManagerSubscriber';
+export { default as SetsManager } from './SetsManager';
